@@ -1,7 +1,7 @@
 package net.andresbustamante.myproject.api.entities;
 
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
 
 public class VersionEntityListener {
 
